finished team 4 activity
