+++ src/js/main.js
@@ -1,3 +1,5 @@
 import productList from "./productList.mjs";
+import loadHeaderFooter from "./utils.mjs";
 
+loadHeaderFooter();
 productList(".product-list", "tents");
